switch to double values in color

+++ JPEG-Encoder/encoder.console/lib/Color.cs
@@ -2,15 +2,15 @@
 {
   public class Color
   {
-    public Color(int channel1, int channel2, int channel3)
+    public Color(double channel1, double channel2, double channel3)
     {
       Channel1 = channel1;
       Channel2 = channel2;
       Channel3 = channel3;
     }
 
-    public int Channel1 { get; set; }
-    public int Channel2 { get; set; }
-    public int Channel3 { get; set; }
+    public double Channel1 { get; set; }
+    public double Channel2 { get; set; }
+    public double Channel3 { get; set; }
   }
 }
